pyproject.toml
[tool.poetry]
name = "async-python-sprint-4"
version = "0.1.0"
description = ""
authors = []
readme = "README.md"
packages = []

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.86.0"
uvicorn = "^0.19.0"
orjson = "^3.8.1"
tortoise-orm = "^0.19.2"
asyncpg = "^0.27.0"
fastapi-pagination = "^0.10.0"


[tool.poetry.group.dev.dependencies]
pytest = "^7.2.0"
pytest-asyncio = "^0.20.1"
requests = "^2.28.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
